Refactor of guideline logic. Old `BaseGuideline` required implementing classes to provide a static error code and a `validate` function. If the validation function returns true then the error code is raised and used.
This pattern with a single static error code per guideline implementation makes it hard to implement rules that has dependencies on each other. It also requires many different rules to validate a single property in order to be able to signal different error codes.

This logic is changed. The `validate` function will now return the error code (if any, else null) to raise if there are any problems.